Fixing all tests
Next step is to set up travis and the gem up-to-date check.
After that, improve the code itself. Everything related to registration
periods is a bit messy as it has an event associated but doesn't filter
by it yet.
Oh, and, of course, OAuth is mandatory so we can start providing admin
access and similars.