Fixing notification comming from BCash.
BCash doesn't respect their website's description of the notification post.
They do not provide any of the information about the payment except for the
id of the transaction and the id of the sale.
It seems like the expected implementation is to store some data upon callback
and just change that data upon payment notification.
This only fixes so that a payment notification is enough. Created #21 to
handle this issue a little better.