Closes #53
This adds a column for the elements that are not pending but also have
no payment associated to them. Those can be manual changes as well as
external payments.
Also improved performance for this page considerably and fixed and error
when trying to add notes to an attendance upon creation.