Fixing test problem
Because of the if session[:user_id] in the current user method, the
current user was not being returned for the tests or in cases where the
session is absent (even if there is a current user set).
Unlikely to ever happen in production but the logic was flawed anyway
and was making the tests harder to run.