Allowing for a devleoper sign in outside of production.
This has been very annoying over time. Although I hate the idea of
having an 'if production', it is better than commenting out code.
Ideally I'd have this as a gem that I would only include in development
but since omniauth already has it, the if (in two places) will have to do.