Fixing puppet runs for passenger-install-apache2-module
Finally. Thanks to vagrant and some brain power, I figured out that
previously the packages were just being installed out of luck in the
order we needed them. Added the dependencies to the execution of the
installer.
Also added a manifest variable to the deploy since now I want to be able
to deploy to the vagrant deploy box just like I do for digital ocean or
aws.