Adding vagrant-cachier and librarian-puppet
Also fixing the vagrant dev deploy (thanks to librarian puppet)
Now both vagrant boxes (dev and deploy) as well as digital ocean
droplets can be started from scratch to running with just the
first_deploy script.
Haven't tested on an AWS EC2 instance but there should be no reason for
it to not work in there.