Upgrading minor gem versions
Only couple of majors was pagseguro with a new feature (checkout
transparente) and netrc with respecting NETRC environment variable
Omniauth-oauth2 also removed a callback method that it was overriding.
From the documentation it doesn't look like it will affect us in any way
but this is worth keeping an eye on if people have a hard time logging
in.